laird: ath6kl: Move reset GPIO into init
Originally in probe, the reset GPIO would only be grabbed if the chip were
already out of reset and communicating on the SDIO bus. If not the case, the
GPIO wouldn't be grabbed by the driver to pull the chip out of reset. This
moves it to init and exit so that we actually do this properly and the
driver can pull the chip out of reset.

Additionally ensures we do a reset on bring up, and also forces it back into
reset when we unload the driver.

Bug: 7799