laird: ath6kl: add module parameter to disable the fw debug log
The firmware host interest option register gives us an option to disable
the firmware from sending the debug logs. If we don't disable the debug
logs, we get constant information over the SDIO interface. By disabling the
logging, we reduce that information and can increase throughput in some
cases.

This patch adds a module parameter, disable_fw_dbglog, to do exactly this.
A user would set it to 1 and then when the firmware is loaded it will be
instructed to disable the firmware logs via the HI_OPTION_DISABLE_DBGLOG
being set in the host-interest flags.

The existing drivers don't send the flag down and I wanted to keep the
default behavior by having the module parameter specify the disable
instead of explicitly requiring the user to enable the debug. This
maintains consistent expected behavior unless a user expressly chooses to
change the behavior by setting the parameter.

In testing with Laird's WB50, an ar6004 based device, this results in a
basic throughput increase of 1 Mbps, or approximately 3.5% in my test
environment.

Bug: 10410