non-upstream: disable command decode in sdio_aos for 4373
AOS is a part of the SDIOD core that becomes active when the rest of
SDIOD is sleeping to keep SDIO bus alive responding to reduced set of
commands.

Transaction between AOS and SDIOD is not protected, and if cmd 52 is
received in AOS and in the middle of response state changed from AOS to
SDIOD, response is corrupted and it causes to SDIO Host controller to
hang.

Signed-off-by: Chi-hsien Lin <chi-hsien.lin@cypress.com>
(cherry picked from commit 72c9f220374f68970e03c895457bb950050768d3)