brcmfmac: reserve tx credit only when txctl is ready to send
commit 9a2a0862d973c9eab6a1f3efa0d4d18dcedbaf1a upstream.

The 4329 throughput drops from 40.2 Mbits/sec to 190 Kbits/sec in 2G
11n mode because the commit b41c232d33666 ("brcmfmac: reserve 2 credits
for host tx control path"). To fix the issue, host driver only reserves
tx control credit when there is a txctl frame is pending to send. And
we also check available credit by using "not equal to 0" instead of
"greater than 0" because tx_max and tx_seq are circled positive numbers.

Fixes: b41c232d33666 ("brcmfmac: reserve 2 credits for host tx control path")

Signed-off-by: Wright Feng <wright.feng@cypress.com>
Tested-by: Dmitry Osipenko <digetx@gmail.com>